[Bronze II] Title: Hashing, Time: 124 ms, Memory: 14148 KB -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -9,7 +9,7 @@ public class Main{
         long sum = 0;
         for(int i = 0; i < N; i++){
             int tmp = br.read();
-            sum += (tmp - 96) * (Math.pow(31, i));
+            sum += (tmp - 96) * ((long)Math.pow(31, i));
         }
         System.out.println(sum);
     }
